Allow hook to work without disrupting `sails lift`
The hook works great on environments that use `node app.js` to start the application, but causes breaking errors when using `sails lift`. I'm seeing this when developing locally vs on production.

Adding this simple check for the existence of the `sails.newRelic` object will prevent errors when using `sails.lift`.